Rename the component DMS to DM
