+++ components/dm/command/check.go
