+++ components/dm/command/destroy.go
