+++ components/dm/command/restart.go
