+++ components/dm/command/scale_out.go
