+++ components/dm/command/stop.go
